backend/requirements.txt
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.29.0,<1.0
pydantic>=2.7.0,<3.0
pydantic-settings>=2.2.0,<3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.0.0,<3.0
numpy>=1.24.0,<2.0
joblib>=1.3.0
websockets>=12.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0,<3.0
psycopg2-binary>=2.9.11,<3.0

# MQTT
paho-mqtt>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
